Moves methods from helper to api client. Spec file is split up into helper and api client spec files
